python-worker/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
numpy
protobuf
Pillow
onnxruntime
opencv-python-headless